+++ mm/kasan/Make.defs
@@ -18,21 +18,18 @@
 #
 ############################################################################
 
-ifeq ($(CONFIG_MM_KASAN),y)
+CSRCS += hook.c
 
-CSRCS += kasan.c
+ifeq ($(CONFIG_MM_KASAN),y)
+CSRCS += generic.c
 
 # Disable kernel-address in mm subsystem
 
 CFLAGS += -fno-sanitize=kernel-address
-
-ifeq ($(CONFIG_LTO_NONE),n)
-  CFLAGS += -fno-lto
+CFLAGS += -fno-lto
 endif
 
 # Add the core heap directory to the build
 
 DEPPATH += --dep-path kasan
 VPATH += :kasan
-
-endif
