+++ mm/kasan/generic.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * mm/kasan/kasan.c
+ * mm/kasan/generic.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -25,8 +25,6 @@
 #include <nuttx/spinlock.h>
 
 #include <assert.h>
-#include <debug.h>
-#include <stdbool.h>
 #include <stdint.h>
 
 #include "kasan.h"
@@ -60,7 +58,7 @@
 
 #endif
 
-#define KASAN_INIT_VALUE            0xDEADCAFE
+#define KASAN_INIT_VALUE            0xdeadcafe
 
 /****************************************************************************
  * Private Types
@@ -100,7 +98,7 @@ static FAR uintptr_t *kasan_mem_to_shadow(FAR const void *ptr, size_t size,
   FAR struct kasan_region_s *region;
   uintptr_t addr = (uintptr_t)ptr;
 
-  if (g_region_init != KASAN_INIT_VALUE)
+  if (size == 0 || g_region_init != KASAN_INIT_VALUE)
     {
       return NULL;
     }
@@ -136,33 +134,6 @@ static FAR uintptr_t *kasan_mem_to_shadow(FAR const void *ptr, size_t size,
   return NULL;
 }
 
-static void kasan_report(FAR const void *addr, size_t size,
-                         bool is_write,
-                         FAR void *return_address)
-{
-  static int recursion;
-
-  if (++recursion == 1)
-    {
-      _alert("kasan detected a %s access error, address at %p,"
-             "size is %zu, return address: %p\n",
-             is_write ? "write" : "read",
-             addr, size, return_address);
-      PANIC();
-    }
-
-  --recursion;
-}
-
-static bool kasan_is_poisoned(FAR const void *addr, size_t size)
-{
-  FAR uintptr_t *p;
-  unsigned int bit;
-
-  p = kasan_mem_to_shadow(addr + size - 1, 1, &bit);
-  return p && ((*p >> bit) & 1);
-}
-
 static void kasan_set_poison(FAR const void *addr, size_t size,
                              bool poisoned)
 {
@@ -172,15 +143,17 @@ static void kasan_set_poison(FAR const void *addr, size_t size,
   uintptr_t mask;
   int flags;
 
-  flags = spin_lock_irqsave(&g_lock);
-
   p = kasan_mem_to_shadow(addr, size, &bit);
-  DEBUGASSERT(p != NULL);
+  if (p == NULL)
+    {
+      return;
+    }
 
   nbit = KASAN_BITS_PER_WORD - bit % KASAN_BITS_PER_WORD;
   mask = KASAN_FIRST_WORD_MASK(bit);
-
   size /= KASAN_SHADOW_SCALE;
+
+  flags = spin_lock_irqsave(&g_lock);
   while (size >= nbit)
     {
       if (poisoned)
@@ -215,21 +188,18 @@ static void kasan_set_poison(FAR const void *addr, size_t size,
   spin_unlock_irqrestore(&g_lock, flags);
 }
 
-static inline void kasan_check_report(FAR const void *addr, size_t size,
-                                      bool is_write,
-                                      FAR void *return_address)
-{
-  if (kasan_is_poisoned(addr, size))
-    {
-      kasan_report(addr, size, is_write, return_address);
-    }
-}
-
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
-/* Exported functions called from other mm module */
+bool kasan_is_poisoned(FAR const void *addr, size_t size)
+{
+  FAR uintptr_t *p;
+  unsigned int bit;
+
+  p = kasan_mem_to_shadow(addr, size, &bit);
+  return p && ((*p >> bit) & 1);
+}
 
 void kasan_poison(FAR const void *addr, size_t size)
 {
@@ -244,119 +214,21 @@ void kasan_unpoison(FAR const void *addr, size_t size)
 void kasan_register(FAR void *addr, FAR size_t *size)
 {
   FAR struct kasan_region_s *region;
-  int flags;
 
   region = (FAR struct kasan_region_s *)
     ((FAR char *)addr + *size - KASAN_REGION_SIZE(*size));
 
   region->begin = (uintptr_t)addr;
   region->end   = region->begin + *size;
-
-  flags = spin_lock_irqsave(&g_lock);
   region->next  = g_region;
   g_region      = region;
   g_region_init = KASAN_INIT_VALUE;
-  spin_unlock_irqrestore(&g_lock, flags);
 
   kasan_poison(addr, *size);
   *size -= KASAN_REGION_SIZE(*size);
 }
 
-/* Exported functions called from the compiler generated code */
-
-void __sanitizer_annotate_contiguous_container(FAR const void *beg,
-                                               FAR const void *end,
-                                               FAR const void *old_mid,
-                                               FAR const void *new_mid)
+void kasan_init_early(void)
 {
-  /* Shut up compiler complaints */
+  g_region_init = 0;
 }
-
-void __asan_before_dynamic_init(FAR const void *module_name)
-{
-  /* Shut up compiler complaints */
-}
-
-void __asan_after_dynamic_init(void)
-{
-  /* Shut up compiler complaints */
-}
-
-void __asan_handle_no_return(void)
-{
-  /* Shut up compiler complaints */
-}
-
-void __asan_report_load_n_noabort(FAR void *addr, size_t size)
-{
-  kasan_report(addr, size, false, return_address(0));
-}
-
-void __asan_report_store_n_noabort(FAR void *addr, size_t size)
-{
-  kasan_report(addr, size, true, return_address(0));
-}
-
-void __asan_loadN_noabort(FAR void *addr, size_t size)
-{
-  kasan_check_report(addr, size, false, return_address(0));
-}
-
-void __asan_storeN_noabort(FAR void * addr, size_t size)
-{
-  kasan_check_report(addr, size, true, return_address(0));
-}
-
-void __asan_loadN(FAR void *addr, size_t size)
-{
-  kasan_check_report(addr, size, false, return_address(0));
-}
-
-void __asan_storeN(FAR void *addr, size_t size)
-{
-  kasan_check_report(addr, size, true, return_address(0));
-}
-
-#define DEFINE_ASAN_LOAD_STORE(size) \
-  void __asan_report_load##size##_noabort(FAR void *addr) \
-  { \
-    kasan_report(addr, size, false, return_address(0)); \
-  } \
-  void __asan_report_store##size##_noabort(FAR void *addr) \
-  { \
-    kasan_report(addr, size, true, return_address(0)); \
-  } \
-  void __asan_load##size##_noabort(FAR void *addr) \
-  { \
-    kasan_check_report(addr, size, false, return_address(0)); \
-  } \
-  void __asan_store##size##_noabort(FAR void *addr) \
-  { \
-    kasan_check_report(addr, size, true, return_address(0)); \
-  } \
-  void __asan_load##size(FAR void *addr) \
-  { \
-    kasan_check_report(addr, size, false, return_address(0)); \
-  } \
-  void __asan_store##size(FAR void *addr) \
-  { \
-    kasan_check_report(addr, size, true, return_address(0)); \
-  }
-
-DEFINE_ASAN_LOAD_STORE(1)
-DEFINE_ASAN_LOAD_STORE(2)
-DEFINE_ASAN_LOAD_STORE(4)
-DEFINE_ASAN_LOAD_STORE(8)
-DEFINE_ASAN_LOAD_STORE(16)
-
-#ifdef CONFIG_MM_KASAN_GLOBAL
-void __asan_register_globals(void *ptr, ssize_t size)
-{
-  /* Shut up compiler complaints */
-}
-
-void __asan_unregister_globals(void *ptr, ssize_t size)
-{
-  /* Shut up compiler complaints */
-}
-#endif
